Extract away theme loading
